Added DeleteGame, Changed Waitgame scheme
I added deletegame.php, which is intended to be used when a timeout or
exit occurs, to delete that user’s games and lobby queues.  I also
changed the waitgame schema.  Now, there is a DataRead and Uploaded
field, where Uploaded is who last uploaded the most recent data.  This
way, we can track who should be receiving the data, while the client
can track the game state